Move _check_mesh from html_surface into check_mesh in surface.
The reason is use check_mesh also in plot_img_on_surf
