+++ nilearn/plotting/tests/test_html_surface.py
@@ -30,16 +30,6 @@ def test_get_vertexcolor():
     assert len(vertexcolors) == len(mesh[0])
 
 
-def test_check_mesh():
-    mesh = html_surface._check_mesh('fsaverage5')
-    assert mesh is html_surface._check_mesh(mesh)
-    assert_raises(ValueError, html_surface._check_mesh, 'fsaverage3')
-    mesh.pop('pial_left')
-    assert_raises(ValueError, html_surface._check_mesh, mesh)
-    assert_raises(TypeError, html_surface._check_mesh,
-                  surface.load_surf_mesh(mesh['pial_right']))
-
-
 def test_one_mesh_info():
     fsaverage = fetch_surf_fsaverage()
     mesh = surface.load_surf_mesh(fsaverage['pial_right'])
