+++ nilearn/surface/__init__.py
@@ -3,7 +3,8 @@ Functions for surface manipulation.
 """
 
 from .surface import (vol_to_surf, load_surf_data,
-                      load_surf_mesh, check_mesh_and_data)
+                      load_surf_mesh, check_mesh,
+                      check_mesh_and_data)
 
 __all__ = ['vol_to_surf', 'load_surf_data', 'load_surf_mesh',
-           'check_mesh_and_data']
+           'check_mesh', 'check_mesh_and_data']
